requirements.txt
python-telegram-bot
bs4
httpx
requests
google-genai
APScheduler